backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
mysql-connector-python==9.1.0
paho-mqtt==1.6.1
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.4
